DD-376 - Fix re-embedding

Approved-by: Anand Raj

+++ app/common/services/repository/chunk/chunk_service.py
@@ -1,14 +1,14 @@
-import asyncio
-from typing import List
+from datetime import datetime
+from typing import List, Tuple
 
 from sanic.log import logger
 from weaviate.classes.query import Filter
-from weaviate.collections.classes.data import DataObject
 from weaviate.util import generate_uuid5
 
 from app.common.models.dao.weaviate.chunks import Chunks
 from app.common.models.dto.chunk_dto import ChunkDTO, ChunkDTOWithVector
 from app.common.services.repository.dataclasses.main import WeaviateSyncAndAsyncClients
+from app.common.utils.app_logger import AppLogger
 
 
 class ChunkService:
@@ -39,9 +39,9 @@ class ChunkService:
             logger.exception("Failed to get chunk files by commit hashes")
             raise ex
 
-    async def get_chunks_by_chunk_hashes(self, chunk_hashes: List[str]) -> List[ChunkDTO]:
-        BATCH_SIZE = 10000
-        all_chunks = []
+    async def get_chunks_by_chunk_hashes(self, chunk_hashes: List[str]) -> List[Tuple[ChunkDTO, List[float]]]:
+        BATCH_SIZE = 1000
+        all_chunks: List[Tuple[ChunkDTO, List[float]]] = []
         MAX_RESULTS_PER_QUERY = 10000
         try:
             # Process chunk hashes in batches
@@ -49,17 +49,18 @@ class ChunkService:
                 batch_hashes = chunk_hashes[i : i + BATCH_SIZE]
                 batch_chunks = await self.async_collection.query.fetch_objects(
                     filters=Filter.any_of(
-                        [Filter.by_property("chunk_hash").equal(chunk_hash) for chunk_hash in batch_hashes]
+                        [Filter.by_id().equal(generate_uuid5(chunk_hash)) for chunk_hash in batch_hashes]
                     ),
+                    include_vector=True,
                     limit=MAX_RESULTS_PER_QUERY,
                 )
                 # Break if no more results
                 if batch_chunks.objects:
                     # Convert to DTOs efficiently using list comprehension
                     batch_dtos = [
-                        ChunkDTO(**chunk_obj.properties, id=str(chunk_obj.uuid)) for chunk_obj in batch_chunks.objects
+                        (ChunkDTO(**chunk_obj.properties, id=str(chunk_obj.uuid)), chunk_obj.vector["default"])
+                        for chunk_obj in batch_chunks.objects
                     ]
-
                     all_chunks.extend(batch_dtos)
 
             return all_chunks
@@ -72,25 +73,38 @@ class ChunkService:
             raise
 
     async def bulk_insert(self, chunks: List[ChunkDTOWithVector]) -> None:
-        batch_size = 100
-        for i in range(0, len(chunks), batch_size):
-            batch = chunks[i : i + batch_size]
-            await self.async_collection.data.insert_many(
-                [
-                    DataObject(
-                        properties=chunk.dto.model_dump(mode="json", exclude={"id"}),
-                        vector=chunk.vector,
-                        uuid=generate_uuid5(chunk.dto.chunk_hash),
-                    )
-                    for chunk in batch
-                ]
-            )
-            await asyncio.sleep(0.2)
+        with self.sync_collection.batch.dynamic() as _batch:
+            for chunk in chunks:
+                _batch.add_object(
+                    properties=chunk.dto.model_dump(mode="json", exclude={"id"}),
+                    vector=chunk.vector,
+                    uuid=generate_uuid5(chunk.dto.chunk_hash),
+                )
 
-    def cleanup_old_chunks(self, chunk_hashes_to_clean: List[str]) -> None:
-        batch_size = 500
-        for i in range(0, len(chunk_hashes_to_clean), batch_size):
-            batch = chunk_hashes_to_clean[i : i + batch_size]
-            self.sync_collection.data.delete_many(
-                Filter.all_of([Filter.by_id().not_equal(generate_uuid5(chunk_hash)) for chunk_hash in batch])
+    def cleanup_old_chunks(self, last_used_lt: datetime, exclusion_chunk_hashes: List[str]) -> None:
+        batch_size = 1000
+        while True:
+            deletable_objects = self.sync_collection.query.fetch_objects(
+                limit=batch_size,
+                filters=Filter.all_of(
+                    [
+                        *[
+                            Filter.by_id().not_equal(generate_uuid5(chunk_hash))
+                            for chunk_hash in exclusion_chunk_hashes
+                        ],
+                        Filter.by_property("created_at").less_than(last_used_lt),
+                    ]
+                ),
             )
+
+            AppLogger.log_debug(f"{len(deletable_objects.objects)} chunks to be deleted in batch")
+
+            if len(deletable_objects.objects) <= 0:
+                break
+
+            result = self.sync_collection.data.delete_many(
+                Filter.any_of(
+                    [Filter.by_id().equal(obj.uuid) for obj in deletable_objects.objects],
+                )
+            )
+            AppLogger.log_debug(f"chunks deleted. successful - {result.successful}, failed - {result.failed}")
